chore: moved clients to directories
