+++ Clients/Abstraction/IBulletTrainClient.cs
