+++ Clients/Abstraction/IBulletTrainHttpClient.cs
